Fix Slow Login on a notebook in long running app after sleeping #267
